Time: 0 ms (100.00%) | Memory: 18.9 MB (19.39%) - LeetSync

+++ 268-missing-number/missing-number.py
@@ -4,7 +4,7 @@ class Solution:
         n = len(nums) + 1
         max_num = len(nums)
 
-        exp_sum = (0 + max_num) * (n / 2)
+        exp_sum = int((0 + max_num) * (n / 2))
         act_sum = sum(nums)
 
-        return int(exp_sum - act_sum)
+        return exp_sum - act_sum
